pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pms-mock"
version = "0.1.0"
description = "Hotel Property Management System — FastAPI boilerplate"
requires-python = ">=3.12"
dependencies = [
    # Web framework (includes uvicorn, python-multipart, httpx)
    "fastapi[standard]>=0.115.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Auth
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    # Config
    "pydantic-settings>=2.6.0",
    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.1",
    "faker>=30.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "sqlalchemy[mypy]>=2.0.36",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # FastAPI Depends() calls are safe as default args

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"
